Tests/LibWeb/Text/expected/HTML/session-storage-event-fired-to-lazy-window.txt

key: my-key
oldValue: null
newValue: hello
storageArea: sessionStorage
